treewide: Make multiple runqueue task counters 32-bit
Creating 2**32 tasks is impossible due to futex pid limits and wasteful
anyway. Nobody has done it.

If total number of tasks can't exceed 2**32 (and less due to futex pid
limits), then per-runqueue counters can't as well.

This patchset has been sponsored by REX Prefix Eradication Society.

* sched: Make nr_running() return 32-bit value
* sched: Make nr_iowait() return 32-bit value
* sched: Make nr_iowait_cpu() return 32-bit value
* sched: Make multiple runqueue task counters 32-bit

Link: https://lore.kernel.org/r/20210422200228.1423391-1-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-2-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-3-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-4-adobriyan@gmail.com

+++ drivers/cpuidle/governors/menu.c
@@ -136,7 +136,7 @@ static inline int get_loadavg(unsigned long load)
 	return LOAD_INT(load) * 10 + LOAD_FRAC(load) / 10;
 }
 
-static inline int which_bucket(unsigned int duration, unsigned long nr_iowaiters)
+static inline int which_bucket(unsigned int duration, unsigned int nr_iowaiters)
 {
 	int bucket = 0;
 
@@ -169,7 +169,7 @@ static inline int which_bucket(unsigned int duration, unsigned long nr_iowaiters
  * to be, the higher this multiplier, and thus the higher
  * the barrier to go to an expensive C state.
  */
-static inline int performance_multiplier(unsigned long nr_iowaiters, unsigned long load)
+static inline int performance_multiplier(unsigned int nr_iowaiters, unsigned long load)
 {
 	int mult = 1;
 
@@ -291,7 +291,7 @@ static int menu_select(struct cpuidle_driver *drv, struct cpuidle_device *dev,
 	int idx;
 	unsigned int interactivity_req;
 	unsigned int expected_interval;
-	unsigned long nr_iowaiters, cpu_load;
+	unsigned int nr_iowaiters, cpu_load;
 	ktime_t delta_next;
 
 	if (data->needs_update) {
